workspace/tools/media-pipeline/requirements.txt
faster-whisper
ctranslate2
huggingface-hub
python-docx
